requirements-dev.txt
# ───────── Core micro-service ─────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ───────── Retrieval stack ────────────
sentence-transformers>=2.6.0,<2.8.0     # 2.7.0 is latest on PyPI
faiss-cpu==1.8.0                         # binary wheels for cp38-cp311
numpy==1.26.4

# ───────── Scraping & utilities ───────
requests==2.32.2
beautifulsoup4==4.12.3
tqdm==4.66.4
python-dotenv==1.0.1

# Optional: JS-rendered crawling
playwright==1.44.0        # comment out if you don’t need it
nest-asyncio==1.6.0

# ───────── Observability ──────────────
prometheus-client==0.20.0

# ───────── Typing / data models ───────
pydantic==2.7.1
typing-extensions==4.12.1

nest_asyncio

pytest==8.2.0
ruff==0.4.3            # linter / formatter
mypy==1.10.0
pytest-asyncio==0.23.6
httpx==0.27.0          # test client for FastAPI
pre-commit==3.7.0